StudentManagementSystem100052K/src/studentmanagementsystem/SimpleStudentRepository.java
/**
 * Programming Challenge 1 - Final Project
 * @author Bandara B.M.C.K 100052K
 */
package studentmanagementsystem;

import java.util.*;

public class SimpleStudentRepository implements StudentRepository {

    private Map<Long, Student> studentsDb = new HashMap<Long, Student>();

    // overide the implemented methods
    @Override
    public void saveStudent(Student stu) {
        studentsDb.put(stu.getRegNumber(), stu);
    }

    @Override
    public void deleteStudent(Student stu) {
        studentsDb.remove(stu.getRegNumber());
    }

    @Override
    public Student findStudent(long regNumber) {
        Student st = studentsDb.get(regNumber);

        return st;
    }

    @Override
    public void updateStudent(Student stu) {
        studentsDb.put(stu.getRegNumber(), stu);
    }

    @Override
    public List<Student> findAllStudents() {
        List<Student> list = new ArrayList<Student>();

        // iterate trough map and get the names to a list
        Iterator it = studentsDb.entrySet().iterator();
        while (it.hasNext()) {
            Map.Entry pairs = (Map.Entry) it.next();
            list.add((Student) pairs.getValue());
        }
        return list;
    }
}
